services/ai-services/requirements.txt
fastapi==0.104.1
strawberry-graphql==0.214.0
uvicorn[standard]==0.24.0
PyJWT==2.8.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2